chore: :truck: folders internal restructuration
