+++ core/entity/errors.go
