+++ utils/entity/test/id_test.go
@@ -3,7 +3,7 @@ package test
 import (
 	"testing"
 
-	"github.com/renatospaka/transaction/pkg/entity"
+	"github.com/renatospaka/transaction/utils/entity"
 	"github.com/stretchr/testify/assert"
 )
