+++ utils/entity/trail.go
